making the mappings scrapper use the no duo sign in
