+++ backend/web_scraping/professor_name_mappings/index.js
@@ -6,14 +6,14 @@ import {
 import puppeteer from "puppeteer";
 import fs from "fs";
 import path from "path";
-import { authenticate } from "./utils/authentication.js";
-import { getAndProcessNameMappings } from "./utils/workday.js";
+import { workdayLogin } from "../utils/authenticate.js";
+import { goToCourseSectionsPage } from "../utils/course_sections.js";
+import { parseCourseSectionsXlsx } from "./utils/xlsx_to_csv.js";
 
-export const FIND_COURSE_SECTIONS_URL =
-  "https://www.myworkday.com/scu/d/task/1422$3915.htmld";
 export const REQUEST_INTERVAL_MS = 50;
 export const REQUEST_MAX_RETRIES = 1;
 export const browser = await puppeteer.launch({
+  headless: false,
   args: ["--incognito"],
 });
 export const page = await browser.newPage();
@@ -31,11 +31,12 @@ const s3 =
 
 export default async function main() {
   console.log(
-    `Attempting to update professor name mappings with data from term: ${process.env.ACADEMIC_PERIOD}`
+    `Attempting to update professor name mappings with data from the current term`
   );
   try {
     await initDirectoriesAndFiles();
-    await authenticate(process.env.SCU_USERNAME, process.env.SCU_PASSWORD);
+    await workdayLogin(page, process.env.WORKDAY_USERNAME, process.env.WORKDAY_PASSWORD);
+    await goToCourseSectionsPage(page);
     await getAndProcessNameMappings();
     await writeMappings();
   } catch (e) {
@@ -44,6 +45,40 @@ export default async function main() {
     browser.close();
   }
 }
+
+async function getAndProcessNameMappings() {
+  // Download and parse the course sections spreadsheet.
+  await page.waitForSelector("[title='Export to Excel']");
+  const exportToExcelButton = await page.$('[title="Export to Excel"]');
+  await exportToExcelButton.click();
+  await page.waitForSelector("button[title='Download']");
+  const downloadButton = await page.$('button[title="Download"]');
+  const client = await browser.target().createCDPSession();
+  const curDirectory = process.cwd();
+  await client.send("Browser.setDownloadBehavior", {
+    behavior: "allow",
+    downloadPath: curDirectory,
+    eventsEnabled: true,
+  });
+  await downloadButton.click();
+  await waitForDownload(client);
+  parseCourseSectionsXlsx(`${curDirectory}/SCU_Find_Course_Sections.xlsx`);
+
+  // Clean up the downloaded file.
+  fs.rmSync(`${curDirectory}/SCU_Find_Course_Sections.xlsx`);
+}
+
+async function waitForDownload(session) {
+  return new Promise((resolve) => {
+    session.on("Browser.downloadProgress", (e) => {
+      if (e.state === "completed") {
+        console.log("Successfully downloaded course section spreadsheet.");
+        resolve();
+      }
+    });
+  });
+}
+  
 
 async function initDirectoriesAndFiles() {
   // Running locally. Check if the files exist, and if not, create them.
